refactory back-end structure
