+++ back-end/routes/index.js
@@ -1,0 +1,7 @@
+const express  = require('express');
+const router   = express.Router();
+const v1Router = require('./v1.js');
+
+router.use('/v1', v1Router);
+
+module.exports = router;
