+++ back-end/setup.txt
@@ -6,9 +6,11 @@
 //      1)  npm init -y
 //      2)  npm i express
 //      3)  npm i mysql
+//      4)  npm i promise-sql
 //      4)  npm i express-validator
 //      5)  npm i nodemon --save-dev  
 //      6)  npm i node-fetch
+//      7)  npm i dotenv
 //
 //      API https://www.zipcodeapi.com/API#radiusSql
 //      API https://www.zipcodeapi.com/API#zipToLoc
